@@ -4,7 +4,7 @@
   <Import Project="../../build/version.props" />
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0;net7.0;net8.0;net9.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
     <PackageId>OSharp.Authorization.Datas</PackageId>
     <Summary>OSharp 数据权限组件</Summary>
     <Description>OSharp 数据权限组件，对应用中数据权限进行授权的设计实现</Description>
