@@ -4,7 +4,7 @@
   <Import Project="../../build/version.props" />
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0;net7.0;net8.0;net9.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
     <PackageId>OSharp.Authorization.Functions</PackageId>
     <Summary>OSharp 功能权限组件</Summary>
     <Description>OSharp 功能权限组件，API功能权限授权的设计实现</Description>
