@@ -4,7 +4,7 @@
   <Import Project="../../build/version.props" />
 
   <PropertyGroup>
-    <TargetFrameworks>net6.0;net7.0;net8.0;net9.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;net9.0;net10.0</TargetFrameworks>
     <PackageId>OSharp.AutoMapper</PackageId>
     <Summary>OSharp AutoMapper 对象映射组件</Summary>
     <Description>OSharp AutoMapper 对象映射组件，封装基于AutoMapper的对象映射实现</Description>
