@@ -7,31 +7,43 @@
 //  <last-date>2020-05-28 15:00</last-date>
 // -----------------------------------------------------------------------
 
+using Microsoft.Extensions.Hosting;
+
+
 namespace OSharp.Wpf.Stylet;
 
 public abstract class ServiceProviderBootstrapper<TRootViewModel> : BootstrapperBase where TRootViewModel : class
 {
-    private object _rootViewModel;
-
-    protected virtual object RootViewModel
-    {
-        get { return _rootViewModel ??= ServiceProvider.GetService(typeof(TRootViewModel)); }
-    }
+    private HostApplicationBuilder _hostBuilder;
+    private IHost _host;
+    private readonly CancellationTokenSource _cancellationTokenSource = new();
+    private TRootViewModel _rootViewModel;
+    protected virtual TRootViewModel RootViewModel => _rootViewModel ??= (TRootViewModel)GetInstance(typeof(TRootViewModel));
 
     protected IServiceProvider ServiceProvider { get; private set; }
 
     /// <summary>
+    /// Called on application startup. This occur after this.Args has been assigned, but before the IoC container has been configured
+    /// </summary>
+    protected override void OnStart()
+    {
+        _hostBuilder = Host.CreateApplicationBuilder();
+        _hostBuilder.Environment.EnvironmentName = Environment.GetEnvironmentVariable("ASPNETCORE_ENVIRONMENT")
+            ?? Environment.GetEnvironmentVariable("DOTNET_ENVIRONMENT") ?? "Production";
+    }
+
+    /// <summary>
     /// Overridden from BootstrapperBase, this sets up the IoC container
     /// </summary>
-    protected sealed override void ConfigureBootstrapper()
+    protected override void ConfigureBootstrapper()
     {
-        IServiceCollection services = new ServiceCollection();
+        _hostBuilder.Services.AddSingleton<IHostApplicationBuilder>(_hostBuilder);
+        DefaultConfigureIoC(_hostBuilder.Services);
+        ConfigureIoC(_hostBuilder.Services);
 
-        // Call DefaultConfigureIoC *after* ConfigureIoIC, so that they can customize builder.Assemblies
-        this.DefaultConfigureIoC(services);
-        this.ConfigureIoC(services);
-
-        ServiceProvider = services.BuildServiceProvider();
+        _host = _hostBuilder.Build();
+        ServiceProvider = _host.Services;
+        _host.StartAsync(_cancellationTokenSource.Token).GetAwaiter().GetResult();
     }
 
     protected virtual void ConfigureIoC(IServiceCollection services)
@@ -39,22 +51,21 @@ public abstract class ServiceProviderBootstrapper<TRootViewModel> : Bootstrapper
 
     protected virtual void DefaultConfigureIoC(IServiceCollection services)
     {
-        ViewManagerConfig viewManagerConfig = new ViewManagerConfig()
+        var viewManagerConfig = new ViewManagerConfig()
         {
-            ViewFactory = this.GetInstance,
-            ViewAssemblies = new List<Assembly>() { GetType().Assembly }
+            ViewFactory = GetInstance,
+            ViewAssemblies = [GetType().Assembly]
         };
 
-        services.AddSingleton<ViewManagerConfig>(viewManagerConfig);
-        services.AddSingleton<IViewManager, ViewManager>();
-        services.AddSingleton<IWindowManagerConfig>(this);
-        services.AddSingleton<IWindowManager>(p => new WindowManager(
-            p.GetRequiredService<IViewManager>(),
-            p.GetRequiredService<IMessageBoxViewModel>,
-            p.GetRequiredService<IWindowManagerConfig>()));
-        services.AddSingleton<IEventAggregator, EventAggregator>();
-        services.AddTransient<IMessageBoxViewModel, MessageBoxViewModel>();
+        services.AddSingleton<IViewManager>(new ViewManager(viewManagerConfig));
         services.AddTransient<MessageBoxView>();
+
+        services.AddSingleton<IWindowManagerConfig>(this);
+        services.AddSingleton<IWindowManager, WindowManager>();
+        services.AddSingleton<IEventAggregator, EventAggregator>();
+        services.AddTransient<IMessageBoxViewModel, MessageBoxViewModel>(); // Not singleton!
+        // Also need a factory
+        services.AddSingleton<Func<IMessageBoxViewModel>>(() => new MessageBoxViewModel());
     }
 
     /// <summary>
@@ -75,12 +86,39 @@ public abstract class ServiceProviderBootstrapper<TRootViewModel> : Bootstrapper
         return ServiceProvider?.GetService(type);
     }
 
+    /// <summary>Hook called on application exit</summary>
+    /// <param name="e">The exit event data</param>
+    protected override void OnExit(ExitEventArgs e)
+    {
+        base.OnExit(e);
+        // 在应用程序退出时停止 host
+        _cancellationTokenSource.Cancel();
+        try
+        {
+            _host?.StopAsync().GetAwaiter().GetResult();
+        }
+        catch (Exception ex)
+        {
+            // 记录日志但不阻止退出
+            Debug.WriteLine($"Error stopping host during exit: {ex.Message}");
+        }
+    }
+
     /// <summary>
     /// Performs application-defined tasks associated with freeing, releasing, or resetting unmanaged resources.
     /// </summary>
     public override void Dispose()
     {
         base.Dispose();
+        try
+        {
+            _host?.StopAsync().GetAwaiter().GetResult();
+        }
+        finally
+        {
+            _host?.Dispose();
+            _cancellationTokenSource.Dispose();
+        }
         ScreenExtensions.TryDispose(_rootViewModel);
     }
 }
