@@ -1,78 +1,57 @@
-﻿// -----------------------------------------------------------------------
-//  <copyright file="BatchingLogger.cs" company="OSharp开源团队">
-//      Copyright (c) 2014-2017 OSharp. All rights reserved.
-//  </copyright>
-//  <site>http://www.osharp.org</site>
-//  <last-editor></last-editor>
-//  <last-date>2017-09-17 21:10</last-date>
-// -----------------------------------------------------------------------
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in https://github.com/aspnet/Logging for license information.
+// https://github.com/aspnet/Logging/blob/2d2f31968229eddb57b6ba3d34696ef366a6c71b/src/Microsoft.Extensions.Logging.AzureAppServices/Internal/BatchingLogger.cs
 
 using System;
 using System.Text;
 
 using Microsoft.Extensions.Logging;
 
+using OSharp.Logging.RollingFile.Formatters;
+
 
 namespace OSharp.Logging.RollingFile.Internal
 {
-//power by https://github.com/andrewlock/NetEscapades.Extensions.Logging
-    internal class BatchingLogger : ILogger
+    public class BatchingLogger : ILogger
     {
-        private readonly string _category;
         private readonly BatchingLoggerProvider _provider;
+        private readonly string _category;
+        private readonly ILogFormatter _formatter;
 
-        public BatchingLogger(BatchingLoggerProvider loggerProvider, string categoryName)
+        public BatchingLogger(BatchingLoggerProvider loggerProvider, string categoryName, ILogFormatter formatter)
         {
             _provider = loggerProvider;
             _category = categoryName;
+            _formatter = formatter;
         }
 
         public IDisposable BeginScope<TState>(TState state)
         {
-            return null;
+            // NOTE: Differs from source
+            return _provider.ScopeProvider?.Push(state);
         }
 
         public bool IsEnabled(LogLevel logLevel)
         {
-            if (logLevel == LogLevel.None)
-            {
-                return false;
-            }
-            return true;
+            return _provider.IsEnabled;
         }
 
-        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
-        {
-            Log(DateTimeOffset.Now, logLevel, eventId, state, exception, formatter);
-        }
-
-        public void Log<TState>(DateTimeOffset timestamp,
-            LogLevel logLevel,
-            EventId eventId,
-            TState state,
-            Exception exception,
-            Func<TState, Exception, string> formatter)
+        public void Log<TState>(DateTimeOffset timestamp, LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
         {
             if (!IsEnabled(logLevel))
             {
                 return;
             }
 
+            var logEntry = new LogEntry<TState>(timestamp, logLevel, _category, eventId, state, exception, formatter);
             var builder = new StringBuilder();
-            builder.Append(timestamp.ToString("yyyy-MM-dd HH:mm:ss.fff zzz"));
-            builder.Append(" [");
-            builder.Append(logLevel.ToString());
-            builder.Append("] ");
-            builder.Append(_category);
-            builder.Append(": ");
-            builder.AppendLine(formatter(state, exception));
-
-            if (exception != null)
-            {
-                builder.AppendLine(exception.ToString());
-            }
-
+            _formatter.Write(in logEntry, _provider.ScopeProvider, builder);
             _provider.AddMessage(timestamp, builder.ToString());
+        }
+
+        public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception exception, Func<TState, Exception, string> formatter)
+        {
+            Log(DateTimeOffset.Now, logLevel, eventId, state, exception, formatter);
         }
     }
 }
